pyproject.toml
[project]
authors = [
    {name = "Muhammad Radifar"},
    {name = "Enade Perdana Istyastono"},
]
license = {text = "Apache License 2.0"}
requires-python = "==3.11.11"
dependencies = [
    "numpy<2.0",
    "rdkit<2024.3",
    "openbabel-wheel<4.0.0,>=3.1.1",
    "torch==2.5.1",
    "biopython<2.0,>=1.84",
    "rich<14.0.0,>=13.9.4",
    "typer<1.0.0,>=0.15.1",
    "meeko",
    "gemmi<1.0.0,>=0.7.0",
    "scipy<2.0.0,>=1.14.1",
]
name = "pilah"
description = "Protein Ligand Abstraction and Hydrogenization tool"
readme = "README.md"
dynamic = ["version"]

[tool.uv.sources]
meeko = { git = "https://github.com/forlilab/Meeko", rev = "4c960993a7b4c0dc2ab3fd8c29e75b30b969bc8a" }
torch = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cpu"

[project.scripts]
pilah = "pilah.pilah:app"
pKAI = "pilah.pKAI.pKAI:main"
dimorphite_dl = "pilah.dimorphite_dl.dimorphite_dl:main"

[dependency-groups]
dev = [
    "jupyterlab>=4.1.6",
    "pre-commit>=4.0.1",
    "pytest-cov>=6.0.0",
]

[build-system]
requires = ["setuptools >= 64.0", "setuptools_scm>=8"]
build-backend = "setuptools.build_meta"

[tool.uv]
default-groups = []
dev-dependencies = [
    "setuptools-scm>=8.1.0",
]

[tool.setuptools_scm]

[tool.ruff]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
    "src/pilah/pKAI",
    "src/pilah/dimorphite_dl"
]

line-length = 119
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "B", "I"]
unfixable = ["B"]
ignore = ["E501", "B006"]

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 119
